+++ voucher-impl/src/main/java/com/dwarfeng/voucher/impl/handler/Pusher.java
@@ -1,0 +1,25 @@
+package com.dwarfeng.voucher.impl.handler;
+
+import com.dwarfeng.subgrade.stack.exception.HandlerException;
+
+/**
+ * 事件推送器。
+ *
+ * @author DwArFeng
+ * @since beta-1.0.0
+ */
+public interface Pusher {
+
+    /**
+     * 返回制造器是否支持指定的类型。
+     *
+     * @param type 指定的类型。
+     * @return 制造器是否支持指定的类型。
+     */
+    boolean supportType(String type);
+
+    /**
+     * 检查重置时执行的推送操作。
+     */
+    void checkReset() throws HandlerException;
+}